+++ lib/snek/board/snake.ex
@@ -20,7 +20,7 @@ defmodule Snek.Board.Snake do
   A valid direction for a snake to move according to the game rules.
   """
   @typedoc since: "0.1.0"
-  @type snake_move :: :north | :south | :east | :west | :forward | :backward | :left | :right
+  @type snake_move :: :up | :down | :left | :right
 
   @typedoc """
   Whether a snake is currently alive, or has been eliminated.
@@ -86,17 +86,8 @@ defmodule Snek.Board.Snake do
   If the snake is already eliminated or the snake does not have any body parts,
   no move will be applied and the snake will remain unchanged.
 
-  If the direction given is `:left`, `:right`, `:forward` or `:backward`, then
-  the snake is moved in that direction relative to the snake's last moved
-  direction.
-
   If the direction given is `nil`, or not a valid direction in which to move,
-  the snake will be moved in the `:forward` direction. If the snake does not
-  have both head and neck body parts, or the head and neck are at teh same
-  position, the snake will default to moving `:north` instead, as in that case
-  the last moved direction cannot be extrapolated.
-
-  If the snake is already eliminated it will not be moved.
+  the snake will be moved in the `:up` direction instead.
 
   Returns the modified snake.
 
@@ -104,7 +95,7 @@ defmodule Snek.Board.Snake do
 
       iex> body = [Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
       iex> snake = %Snake{id: "mysnek", state: :alive, health: 100, body: body}
-      iex> Snake.move(snake, :north)
+      iex> Snake.move(snake, :up)
       %Snake{
         id: "mysnek",
         state: :alive,
@@ -114,31 +105,21 @@ defmodule Snek.Board.Snake do
 
       iex> body = [Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
       iex> snake = %Snake{id: "mysnek", state: {:eliminated, :starvation}, health: 0, body: body}
-      iex> snake == Snake.move(snake, :north)
+      iex> snake == Snake.move(snake, :up)
       true
 
       iex> snake = %Snake{id: "mysnek", state: :alive, health: 100, body: []}
-      iex> snake == Snake.move(snake, :north)
+      iex> snake == Snake.move(snake, :up)
       true
 
       iex> body = [Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
       iex> snake = %Snake{id: "mysnek", state: :alive, health: 100, body: body}
-      iex> snake |> Snake.move(:east) |> Snake.move(nil)
+      iex> snake |> Snake.move(:right) |> Snake.move(nil)
       %Snake{
         id: "mysnek",
         state: :alive,
         health: 100,
-        body: [Snek.Board.Point.new(3, 1), Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1)]
-      }
-
-      iex> body = [Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "mysnek", state: :alive, health: 100, body: body}
-      iex> snake |> Snake.move(:east) |> Snake.move(:left) |> Snake.move(:right)
-      %Snake{
-        id: "mysnek",
-        state: :alive,
-        health: 100,
-        body: [Snek.Board.Point.new(3, 0), Snek.Board.Point.new(2, 0), Snek.Board.Point.new(2, 1)]
+        body: [Snek.Board.Point.new(2, 0), Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1)]
       }
 
   """
@@ -154,20 +135,12 @@ defmodule Snek.Board.Snake do
     snake
   end
 
-  def move(%Snake{body: [head | _rest]} = snake, direction) when direction in [:north, :south, :east, :west] do
+  def move(%Snake{body: [head | _rest]} = snake, direction) when direction in [:up, :down, :left, :right] do
     slither(snake, Point.step(head, direction))
   end
 
-  def move(%Snake{body: [head, neck | _rest]} = snake, direction) when direction in [:forward, :backward, :left, :right] and head != neck do
-    slither(snake, step(snake, direction))
-  end
-
-  def move(%Snake{body: [head, neck | _rest]} = snake, nil) when head != neck do
-    slither(snake, step(snake, :forward))
-  end
-
   def move(%Snake{body: [head | _rest]} = snake, _direction) do
-    slither(snake, Point.step(head, :north))
+    slither(snake, Point.step(head, :up))
   end
 
   defp slither(snake, new_head) do
@@ -181,91 +154,35 @@ defmodule Snek.Board.Snake do
 
   If the snake has no body parts, `nil` is returned instead of a point.
 
-  If the direction given is `:left`, `:right`, `:forward`, or `:backward` then
-  the point returned will be in that direction relative to the snake's last
-  moved direction. If the snake does not have both head and neck body parts, or
-  the head and neck are at the same position, then `nil` will be returned
-  instead of a point, as in that case the last moved direction cannot be
-  extrapolated.
-
   ## Examples
 
       iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
       iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :north)
+      iex> Snake.step(snake, :up)
       %Snek.Board.Point{x: 2, y: 0}
-
-      iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :left)
-      %Snek.Board.Point{x: 2, y: 0}
-
-      iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :east)
-      %Snek.Board.Point{x: 3, y: 1}
-
-      iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :forward)
-      %Snek.Board.Point{x: 3, y: 1}
 
       iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
       iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
       iex> Snake.step(snake, :right)
+      %Snek.Board.Point{x: 3, y: 1}
+
+      iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
+      iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
+      iex> Snake.step(snake, :down)
       %Snek.Board.Point{x: 2, y: 2}
 
       iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
       iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :south)
-      %Snek.Board.Point{x: 2, y: 2}
-
-      iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :backward)
-      %Snek.Board.Point{x: 1, y: 1}
-
-      iex> body = [Snek.Board.Point.new(2, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "snek0", state: :alive, health: 99, body: body}
-      iex> Snake.step(snake, :west)
+      iex> Snake.step(snake, :left)
       %Snek.Board.Point{x: 1, y: 1}
 
       iex> snake = %Snake{id: "snek0", state: :alive, health: 0, body: []}
-      iex> Snake.step(snake, :south)
-      nil
-
-      iex> body = [Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1), Snek.Board.Point.new(1, 1)]
-      iex> snake = %Snake{id: "snek0", state: :alive, health: 100, body: body}
-      iex> Snake.step(snake, :forward)
+      iex> Snake.step(snake, :down)
       nil
 
   """
   @doc since: "0.2.0"
   @spec step(t, Point.direction | snake_move) :: Point.t
-
-  def step(%Snake{body: [head, neck | _rest]}, :forward) when head != neck do
-    vector = Point.difference(head, neck)
-    Point.sum(head, vector)
-  end
-
-  def step(%Snake{body: [head, neck | _rest]}, :backward) when head != neck do
-    vector = Point.difference(neck, head)
-    Point.sum(head, vector)
-  end
-
-  def step(%Snake{body: [head, neck | _rest]}, :left) when head != neck do
-    vector = Point.difference(head, neck) |> Point.rotate_counterclockwise
-    Point.sum(head, vector)
-  end
-
-  def step(%Snake{body: [head, neck | _rest]}, :right) when head != neck do
-    vector = Point.difference(head, neck) |> Point.rotate_clockwise
-    Point.sum(head, vector)
-  end
-
-  def step(_snake, direction) when direction in [:forward, :backward, :left, :right] do
-    nil
-  end
 
   def step(%Snake{body: [head | _rest]}, direction) when not is_nil(head) do
     Point.step(head, direction)